requirements.txt
absl-py==1.4.0
astunparse==1.6.3
blinker==1.7.0
blis==0.7.10
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
confection==0.1.3
cymem==2.0.7
Cython==3.0.4
Flask==3.0.0
flask-ngrok==0.0.25
flask-swagger-ui==4.11.1
flatbuffers==23.5.26
gast==0.4.0
google-api-core==2.19.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
h5py==3.9.0
idna==3.4
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.14.0
langcodes==3.3.0
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
ml-dtypes==0.2.0
murmurhash==1.0.9
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.76
opt-einsum==3.3.0
packaging==23.1
pandas==2.1.2
pathy==0.10.2
Pillow==10.1.0
preshed==3.0.8
proto-plus==1.23.0
protobuf==4.24.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.12
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.11.3
six==1.16.0
smart-open==6.4.0
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.7
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tflearn==0.5.0
thinc==8.1.12
tk==0.1.0
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
wasabi==1.1.2
Werkzeug==3.0.1
wrapt==1.14.1
